Refactored ProfileScreen
